Add files via upload
Hurst exponent brief explanation and calculation process. The Hurst exponent can be used to show a stock's autocorrelation, or correlation with itself, over any period of time, regarding its price data. This is used to determine if it has followed a random walk trend, a mean reverting trend, or a momentum-following trend. It is calculated by obtaining the standard deviation of the differences between a time series that is lagged(by an arbitrary amount) and the original time series, and then plotting that against 
the lag amount used. The slope of this graph is taken as the Hurst exponent, with potentially different slopes at different lags, as 
the function is polynomial